CARDS-2509: Reference question should support conditions based on source form metadata Cleanup: Remove ConditionalOperand and move functionality to base Operand
